Change app connectivity to RetroFit client
Changed our client from OkHTTP to Retrofit to future proof our code. We used the android workshop as a guide to create the files and connectivity. We also added a button that responds to clicks. The app will, at it's current state, crash due to the fact that it has no car to talk to.

Resolves #31