Change webserver implementation
Changed from using just WebServer.h to this example: https://github.com/DIT112-V20/kotlin-app-arduino-sketch-ci/blob/master/arduino/SmartcarWebserver/SmartcarWebserver.ino
Added in some logic for handling requests such as /random and /dance?id=1. Also changed some comments and moved the random dance move logic into a separate method for now as we're thinking of randomizing dances in the app logic and not on the car itself. As it is now, the random dance will loop 5 times which means that 5 moves will be performed by the car.

The mDNS name of the car is set to be djsmartcar and should be accessible through djsmartcar.local.

Resolves #30