Obstacle avoidance within dance moves
Commented out some Serial.print's and moved some magic numbers for speed. Prepared some functionality for requests with a speed parameter that should be able to set the speed of the car. Created a method called ObstacleAvoidance that is responsbile for making sure the car does not run into things while dancing. The method is called in the dancing loops.

Resolves #16