Update button visibility
Updated the visibility of the start and stop button so they appear when a dance move has been chosen. Changed the background color of the home button so it matches the background color of the app. Added in slashes in front of dance and random in the endpoint file so we don't have to add it in everytime we use a mockserver link.

Resolves #44