Change obstacle avoidance speed
Changed the speed set after obstacle avoidance to be one of the 'magic number speed's'. Seems to work better but it was reported that it sometimes acts weirdly during two-step.

Changed the order of the variables at the top of the file and made sure they were following the same naming conventions. Removed a variable called header because it seemed to be unused. Removed some outdated comments.

Resolves #16