Create dance loop
Moved wifi server setup to a separate method instead of having it in setup(). Created variables with ID's for each dance which will help with keeping track of which dance is active. Now calling handleInput() from the loop() using activeDanceId which is set when the server receives a request. Added '/stop'.

Resolves #71