Remove 'stop' and 'random' requests
Car:
Removed randomDance() and the handling of stop and random requests. Made it possible to pass speed and delay to the car in parameters in '/dance' requests. The default is speed = 30 and ms_delay = 1000. Removed all traces of startingPoint in the dance moves. Uncommented Serial.print's to debug twoStep().

App:
Added in parameters for a longer query in Endpoint. Added speed and delay to Dance. Removed stop() and getRandom() methods. Moved some stuff around in MainActivity.kt. Added a recursive call in getDance().

Resolves #71, #88