Enable app to get tempo from Spotify API
Added 'track not found' string to values document.
Added Coroutines implementation to build.gradle.
Added java classes AudioAnalysis and Meta that are needed for the Track class to work properly.
Changed the getTrackAnalysis function in Endpoint to properly take in the Authentification Token and the track ID number.
Added functions to SpotifyService to request an authorisation token from the API and then use it to access the track analysis for the track tempo. Once the API call is finished, the tempo is updated.

Resolves #81.